Title: 계단 오르기, Time: 198ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 231114/계단 오르기/climbing-stairs.java
@@ -4,8 +4,8 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
 
-        int[] arr = new int[n+1];
-        if(n < 3) arr = new int[4];
+        long[] arr = new long[n+1];
+        if(n < 3) arr = new long[4];
         arr[2] = arr[3] = 1;
         for(int i=4; i<=n; i++){
             arr[i] = arr[i-2] + arr[i-3];
